==> code/1_sincronizar_github.R <==
# Configurar Github


# 1. Git ------------------------------------------------------------------

usethis::use_git()


# 2. Github ---------------------------------------------------------------

usethis::use_github(private = T)
